requirements.txt
fastapi
uvicorn[standard]
pymongo
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
ai21
chromadb
argcomplete==3.6.2
asttokens==2.4.1
attrs==25.1.0
Automat==24.8.1
Beaker==1.12.1
beautifulsoup4==4.13.4
blivet==3.12.1
blivet-gui==2.6.0
boto3==1.38.32
botocore==1.38.32
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.1
comm==0.2.2
constantly==23.10.4
crypt_r==3.13.1
cryptography==44.0.0
cupshelpers==1.0
Cython==3.0.12
dasbus==1.7
dbus-python==1.3.2
dbus_next==0.2.3
debugpy==1.8.11
decorator==5.1.1
deluge==2.2.0
distro==1.9.0
dlib==19.24.8
dnf==4.23.0
evdev==1.9.1
executing==2.2.0
file-magic==0.4.0
GeoIP==1.3.2
git-filter-repo==2.47.0
humanize==4.12.0
hyperlink==21.0.0
i3ipc==2.2.1
idna==3.10
incremental==24.7.2
ipykernel==6.29.5
ipython==8.36.0
iso639==0.1.4
jedi==0.19.2
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langtable==0.0.68
libcomps==0.1.21
libdnf==0.74.0
libtorrent==2.0.11
libvirt-python==11.0.0
lutris==0.5.19
lxml==5.3.2
Mako==1.2.3
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
moddb==0.12.0
mutagen==1.47.0
nest-asyncio==1.6.0
nftables==0.1
nodeenv==1.9.1
numexpr==2.10.2
numpy==2.2.6
nwg-wrapper==0.1.3
olefile==0.47
opencv==4.11.0
packaging==24.2
parso==0.8.4
Paste==3.10.1
perf==0.1
pexpect==4.9.0
pid==2.2.3
pillow==11.1.0
platformdirs==4.3.6
ply==3.11
productmd==1.45
prompt_toolkit==3.0.41
protobuf==3.19.6
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pwquality==1.4.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycryptodomex==3.21.0
pycups==2.0.4
pygame==2.6.1
Pygments==2.18.0
PyGObject==3.50.0
pyinotify==0.9.6
pykickstart==3.62
pyOpenSSL==25.0.0
pyparted==3.13.0
pypresence==4.3.0
pyproj==3.7.0
pyright==1.1.390
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-linux-procfs==0.7.3
python-meh==0.52
python-pam==2.0.2
python-xlib==0.33
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.2
pyynl @ file:///builddir/build/BUILD/kernel-6.14.9-build/kernel-6.14.9/linux-6.14.9-300.fc42.x86_64/tools/net/ynl
pyzmq==26.2.0
referencing==0.36.2
rencode==1.0.6
requests==2.32.3
requests-file==2.0.0
requests-ftp==0.3.1
rpds-py==0.25.0
rpm==4.20.1
s3transfer==0.13.0
selinux @ file:///builddir/build/BUILD/libselinux-3.8-build/libselinux-3.8/src
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.8-build/selinux-3.8/python/sepolicy
service-identity==24.2.0
setools==4.5.1
setroubleshoot @ file:///builddir/build/BUILD/setroubleshoot-3.3.35-build/setroubleshoot-3.3.35/src
setuptools==74.1.3
simpleline==1.9.0
six==1.17.0
sos==4.9.1
soupsieve==2.7
stack_data==0.6.3
systemd-python==235
Tempita==0.5.2
tenacity==8.2.3
tornado==6.4.2
traitlets==5.14.3
Twisted==24.11.0
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
websockets==14.1
xkbregistry==0.3
yt-dlp==2025.5.22
zope.interface==7.2
zstandard==0.23.0
